+++ test/ham_test.rb
@@ -393,7 +393,7 @@ class HAMTest < Minitest::Test
   def test_to_area_returns_rectangle
     coord = HAM.new("FN31pr")
     area = coord.to_area
-    assert_instance_of Geodetic::Areas::Rectangle, area
+    assert_instance_of Geodetic::Areas::BoundingBox, area
   end
 
   def test_to_area_contains_midpoint